Move code related to localStorage to stage.js
