+++ lib/game.js
@@ -5,7 +5,6 @@ const qs = require('querystring');
 const Stage = require('./stage');
 const {wait} = require('./util');
 const {calculateScore} = require('./validator');
-const assert = require('assert');
 
 
 class Game {
@@ -190,81 +189,6 @@ class Game {
 
 	closeResult() {
 		$('.result').hide();
-	}
-
-	saveStage() {
-		const boardData = this.currentStage.board.boardData;
-		const stageName = this.configs[this.stageIndex].name;
-		const timestamp = Date.now();
-		const item = localStorage.getItem('boardData');
-		let properties = [];
-		if (item !== null) {
-			properties = JSON.parse(localStorage.getItem('boardData'));
-		}
-		properties.push({stageName, timestamp, boardData});
-		try {
-			localStorage.setItem('boardData', JSON.stringify(properties));
-			return true;
-		} catch (error) {
-			return false;
-		}
-	}
-
-	getStageIndexFromName(name) {
-		return this.configs.findIndex((config) => (name === config.name));
-	}
-
-	loadStage(stageName, timestamp) {
-		const item = localStorage.getItem('boardData');
-		const properties = JSON.parse(item);
-
-		const index = properties.findIndex((property) => (
-			property.stageName === stageName &&
-			property.timestamp === timestamp
-		));
-
-		if (index === -1) {
-			return;
-		}
-		const boardData = properties[index].boardData;
-
-		if (stageName === this.configs[this.stageIndex].name &&
-				this.currentStage !== null) {
-			this.currentStage.clearBoard();
-		} else {
-			this.exitStage();
-			this.startStage(this.getStageIndexFromName(stageName));
-		}
-		this.currentStage.makeBoard(boardData);
-	}
-
-	getSaveEntries(stageName) {
-		const item = localStorage.getItem('boardData');
-		if (item === null) {
-			return [];
-		}
-		const properties = JSON.parse(item);
-		return properties.filter((prop) => (stageName === prop.stageName));
-	}
-
-	removeSaveEntry(stageName, timestamp) {
-		const item = localStorage.getItem('boardData');
-		let properties = [];
-		if (item !== null) {
-			properties = JSON.parse(localStorage.getItem('boardData'));
-		}
-		const index = properties.findIndex((property) => (
-			property.stageName === stageName &&
-			property.timestamp === timestamp
-		));
-		assert(index !== -1);
-		properties.splice(index, 1);
-		try {
-			localStorage.setItem('boardData', JSON.stringify(properties));
-			return true;
-		} catch (error) {
-			return false;
-		}
 	}
 
 	exitStage() {
