+++ lib/stage.js
@@ -120,7 +120,7 @@ class Stage {
 			}
 		});
 		this.$savePanel.find('button.save-entry').click(() => {
-			this.game.saveStage();
+			this.saveStage();
 			this.updateSaveEntries();
 		});
 
@@ -142,8 +142,76 @@ class Stage {
 		this.updateStyles();
 	}
 
+	saveStage() {
+		const boardData = this.board.boardData;
+		const stageName = this.config.name;
+		const timestamp = Date.now();
+		const item = localStorage.getItem('boardData');
+		let properties = [];
+		if (item !== null) {
+			properties = JSON.parse(localStorage.getItem('boardData'));
+		}
+		properties.push({stageName, timestamp, boardData});
+		try {
+			localStorage.setItem('boardData', JSON.stringify(properties));
+			return true;
+		} catch (error) {
+			return false;
+		}
+	}
+
+	loadStage(stageName, timestamp) {
+		const item = localStorage.getItem('boardData');
+		const properties = JSON.parse(item);
+
+		const index = properties.findIndex((property) => (
+			property.stageName === stageName &&
+			property.timestamp === timestamp
+		));
+
+		if (index === -1) {
+			return;
+		}
+		const boardData = properties[index].boardData;
+		assert(stageName === this.config.name);
+
+		if (this.currentStage !== null) {
+			this.clearBoard();
+		}
+		this.makeBoard(boardData);
+	}
+
+	getSaveEntries(stageName) {
+		const item = localStorage.getItem('boardData');
+		if (item === null) {
+			return [];
+		}
+		const properties = JSON.parse(item);
+		return properties.filter((prop) => (stageName === prop.stageName));
+	}
+
+	removeSaveEntry(stageName, timestamp) {
+		const item = localStorage.getItem('boardData');
+		let properties = [];
+		if (item !== null) {
+			properties = JSON.parse(localStorage.getItem('boardData'));
+		}
+		const index = properties.findIndex((property) => (
+			property.stageName === stageName &&
+			property.timestamp === timestamp
+		));
+		assert(index !== -1);
+		properties.splice(index, 1);
+		try {
+			localStorage.setItem('boardData', JSON.stringify(properties));
+			return true;
+		} catch (error) {
+			return false;
+		}
+	}
+
 	updateSaveEntries() {
-		const entries = this.game.getSaveEntries(this.config.name);
+		const entries = this.getSaveEntries(this.config.name);
 		this.$savePanel.find('.saved-entries').empty().append(entries.map((entry) => (
 			$('<li/>', {
 				class: 'saved-item',
@@ -162,7 +230,7 @@ class Stage {
 			const $timestampButton = $(element);
 			const entry = entries[index];
 			$timestampButton.click(() => {
-				this.game.loadStage(entry.stageName, entry.timestamp);
+				this.loadStage(entry.stageName, entry.timestamp);
 				this.updateSaveEntries();
 			});
 		});
@@ -170,7 +238,7 @@ class Stage {
 			const $delButton = $(element);
 			const entry = entries[index];
 			$delButton.click(() => {
-				this.game.removeSaveEntry(entry.stageName, entry.timestamp);
+				this.removeSaveEntry(entry.stageName, entry.timestamp);
 				this.updateSaveEntries();
 			});
 		});
